backend/requirements.txt
Flask==3.0.0
flask-restx==1.3.0
flask-cors==4.0.0
litellm==1.30.0
python-dotenv==1.0.0
pydantic==2.5.0
Werkzeug==3.0.1
gunicorn==21.2.0 
google-generativeai==0.8.5
SQLAlchemy==2.0.23